requirements.txt
aliyun-python-sdk-alidns==3.0.7
requests==2.31.0
python-dotenv==1.0.0 